Add files via upload
When an error occurs in Internet or Blockchain controls, only the first error will sound the alarm. The alarm will not sound on subsequent errors. After the error is fixed, the duration of the error will be written on the log lines. In summary; The alarm will not play every minute for internet or blockchain controls.