Add files via upload
1. Automatic report sending feature has been added via Microsoft Outlook. Microsoft Outlook (2013, 2016, 2019) application must be installed and account settings must be made on the computer.
Mail format is kept in MailFormat.html file. If there is someone who understands HTML design well, they can edit it.
I'd be happy if you could send it to me too. 😊

2. A daily and weekly comparison report has been added to the Network Health page. You will be able to compare by hour.

3. The time to check availability has been reduced from 6 hours to 30 minutes. And the timeout duration has also been increased. 
For this reason, in case of poor network health, the availability rate control may cause freezes on the screen.
However, it will not check the available rate again on the same day after it is withdrawn.

4. Visual improvements have been made in grid areas.