PiCheck version 6.0.0 is ready.
**** What's new ****
1. PiCheck version 6.0 has been rewritten in a different software tool. No need to install .Net 5.0 package from Microsoft site anymore.
In the new version, boot and process times are even faster. In addition, the amount of RAM used has been significantly reduced.
The "Loading" screen was canceled because the boot was already too fast :)

2. VERY IMPORTANT : The *.log files where the data is kept have been changed because they were rewritten in a different software tool.
"Converting" field has been added at the bottom of the Settings page so that you do not lose your old data. You will only need to do this once.
It will convert the data of the old version to the new version.

3. You do not need to restart the application for the changes you made on the Settings page. Language setting or any other setting change will take effect immediately.
When you change the language setting of the application, the texts on the screen will change instantly.

4. Country names on the Connections page will now appear according to the language you use. If you change the language setting while using the application, the country names will change instantly.

5. Monitor page: Some settings at the bottom have been removed. The Alert setting has been moved to the "Settings" page. Also, warning options have been increased. Added 10 different alarm sounds.

6. Analysis page: A column has been added at the end of the Daily Special field so that you can enter the "Node Bonus" information you receive every day. You may see some fields in the list as bold text. Fields with bold text mean that it is the field with the highest value.

7. Availability page: Node installation date and Availability rate updated date/time information has been added to the list.
Added the Day column. You will be able to compare it to 90 days ago.

8. Connections page: I now define the country names here. In which language you are using the application, you will be able to see the country names in that language.

9. Network Health page : Only 3 ports required for Node Consensus are shown. I removed the other 7 ports.

10. Settings page: I added improvements such as Windows notifications and automatic export. If you minimize the PiCheck application to the system tray, you will be able to receive the errors as Windows Notification.
You will be able to use the application with 4 different theme options. Blue, Dark, Gray and Silver.
Added transparency setting for home screen and widget screen.

11. In the old version, you could only "HTML Export" in grid areas. In PiCheck 6 version, you will be able to export in different formats such as Excel, PDF, Text, XML, CSV, HTML, DBase, Lotus 1-2-3.

12. A bar has been added at the top of the Analysis, Availability, Connections and Network Health pages to show how much time is left for the next action.

13. All date fields are formatted according to the system date format. It is formatted according to your Windows date format.

**** Bugs ****
--When you select a theme and then switch to normal view, the screen title changes to standard Windows format. It does not convert to Windows 10 or Windows 11 graphical interface. Closing and reopening the app fixes the problem.
--The application normally uses 50-70 MB of RAM. However, when the grid areas with too many records are transferred to the Excel file, there is an increase in RAM. But it still continues to use less RAM than the old PiCheck application.

**** How to update? ****
You can copy all the files in the downloaded zip file onto the old PiCheck application.
If you want the old application to remain, you should install the PiCheck 6 version in a different folder.

**** Important Note ****
You may get vcruntime140.dll error when you run the application because some computers do not have the Microsoft Visual C++ package installed.
But this package is usually already installed automatically by many applications. That's why it's very rare to get this error.
I have added this application package to the zip file in case it is still necessary. You can find this package in the "Required" folder.

**** Country Names ****
I translated the country names using Google Translate. There may be errors. Anyone who wants to help with their language can help.
Many thanks to our friend @Jochen for German country names and display texts :)

I hope you enjoy the new version.
Best regards